Fix unsigned short to int conversion in TCPPacket

+++ src/ru/nsu/ccfit/bogush/tou/TOUConnectionManager.java
@@ -255,8 +255,8 @@ class TOUConnectionManager {
     }
 
     private static void swapSourceAndDestination(TCPPacket packet) {
-        short dstPort = packet.sourcePort();
-        short srcPort = packet.destinationPort();
+        int dstPort = packet.sourcePort();
+        int srcPort = packet.destinationPort();
         packet.destinationPort(dstPort);
         packet.sourcePort(srcPort);
     }
